Move io module into tools module
There were several problems with the import when there is an io module in the working directory. To avoid these, the module remains a submodule of tools
